:bug: Fix get testcases api / remove TestcaseRepository
