--- Domain/Problem/src/main/java/tw/waterball/judgegirl/problem/domain/usecases/GetTestCasesUseCase.java
@@ -1,48 +1,0 @@
-/*
- * Copyright 2020 Johnny850807 (Waterball) 潘冠辰
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *       http://www.apache.org/licenses/LICENSE-2.0
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package tw.waterball.judgegirl.problem.domain.usecases;
-
-import lombok.Value;
-import tw.waterball.judgegirl.commons.exceptions.NotFoundException;
-import tw.waterball.judgegirl.primitives.problem.Testcase;
-import tw.waterball.judgegirl.problem.domain.repositories.TestCaseRepository;
-
-import javax.inject.Named;
-import java.util.List;
-
-/**
- * @author - johnny850807@gmail.com (Waterball)
- */
-@Named
-public class GetTestCasesUseCase {
-    private TestCaseRepository testCaseRepository;
-
-    public GetTestCasesUseCase(TestCaseRepository testCaseRepository) {
-        this.testCaseRepository = testCaseRepository;
-    }
-
-    public void execute(Request request, Presenter presenter) throws NotFoundException {
-        List<Testcase> testcases = testCaseRepository.findAllInProblem(request.problemId);
-        presenter.setTestcases(testcases);
-    }
-
-    public interface Presenter {
-        void setTestcases(List<Testcase> testcases);
-    }
-
-    @Value
-    public static class Request {
-        public int problemId;
-    }
-}
