--- Spring-Boot/Spring-Boot-Problem/src/main/java/tw/waterball/judgegirl/springboot/problem/repositories/MongoTestcaseRepository.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2020 Johnny850807 (Waterball) 潘冠辰
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *       http://www.apache.org/licenses/LICENSE-2.0
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
-package tw.waterball.judgegirl.springboot.problem.repositories;
-
-import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.query.Criteria;
-import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.data.mongodb.gridfs.GridFsTemplate;
-import org.springframework.stereotype.Component;
-import tw.waterball.judgegirl.commons.models.files.FileResource;
-import tw.waterball.judgegirl.primitives.problem.Problem;
-import tw.waterball.judgegirl.primitives.problem.Testcase;
-import tw.waterball.judgegirl.problem.domain.repositories.TestCaseRepository;
-import tw.waterball.judgegirl.springboot.profiles.productions.Mongo;
-import tw.waterball.judgegirl.springboot.utils.MongoUtils;
-
-import java.util.List;
-import java.util.Optional;
-
-import static tw.waterball.judgegirl.springboot.utils.MongoUtils.downloadFileResourceByFileId;
-
-/**
- * @author - johnny850807@gmail.com (Waterball)
- */
-@Mongo
-@Component
-public class MongoTestcaseRepository implements TestCaseRepository {
-
-    private final MongoTemplate mongoTemplate;
-    private final GridFsTemplate gridFsTemplate;
-
-    public MongoTestcaseRepository(MongoTemplate mongoTemplate, GridFsTemplate gridFsTemplate) {
-        this.mongoTemplate = mongoTemplate;
-        this.gridFsTemplate = gridFsTemplate;
-    }
-
-    @Override
-    public List<Testcase> findAllInProblem(int problemId) {
-        return mongoTemplate.find(new Query(Criteria.where("problemId").is(problemId)), Testcase.class);
-    }
-
-    @Override
-    public Optional<FileResource> downloadTestCaseIOs(int problemId, String testcaseIOsFileId) {
-        return MongoUtils.query(mongoTemplate)
-                .fromDocument(Problem.class)
-                .selectOneField("testcaseIOsFileId")
-                .byId(problemId)
-                .execute()
-                .getField(Problem::getTestcaseIOsFileId)
-                .map(fileId -> downloadFileResourceByFileId(gridFsTemplate, fileId));
-    }
-}
